requirements.txt
httpx==0.27.0
orjson==3.10.7
pydantic==2.9.2
python-dotenv==1.0.1
networkx==3.2.1
pandas==2.2.2
matplotlib==3.9.2
geoip2==4.8.0
